# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BrandingOut": {
        "properties": {
          "accent_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent Color"
          },
          "font_pair": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Font Pair"
          },
          "footer_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Footer Text"
          },
          "logo_document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Document Id"
          },
          "payment_instructions": {
            "$ref": "#/components/schemas/PaymentDefaults"
          },
          "primary_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Color"
          },
          "templates": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Template key per document type (effective)",
            "title": "Templates",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "templates",
          "primary_color",
          "accent_color",
          "font_pair",
          "logo_document_id",
          "footer_text",
          "payment_instructions",
          "version"
        ],
        "title": "BrandingOut",
        "type": "object"
      },
      "BrandingUpdate": {
        "additionalProperties": false,
        "properties": {
          "accent_color": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent Color"
          },
          "font_pair": {
            "anyOf": [
              {
                "enum": [
                  "sans",
                  "serif"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Font Pair"
          },
          "footer_text": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Footer Text"
          },
          "logo_document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Document Id"
          },
          "payment_instructions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentDefaults"
              },
              {
                "type": "null"
              }
            ]
          },
          "primary_color": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Color"
          },
          "templates": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "enum": [
                    "quote",
                    "invoice",
                    "receipt",
                    "credit_note"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templates"
          }
        },
        "title": "BrandingUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaymentDefaults": {
        "additionalProperties": false,
        "description": "Payment instructions printed on invoices (the per-document reference is added automatically).",
        "properties": {
          "bank_account_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Account Name"
          },
          "bank_account_number": {
            "anyOf": [
              {
                "pattern": "^[0-9A-Za-z -]{4,34}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Account Number"
          },
          "bank_branch": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Branch"
          },
          "bank_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Name"
          },
          "mpesa_paybill": {
            "anyOf": [
              {
                "pattern": "^\\d{5,7}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mpesa Paybill"
          },
          "mpesa_till": {
            "anyOf": [
              {
                "pattern": "^\\d{5,7}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mpesa Till"
          },
          "swift_code": {
            "anyOf": [
              {
                "pattern": "^[A-Z0-9]{8}([A-Z0-9]{3})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swift Code"
          }
        },
        "title": "PaymentDefaults",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__modules__rendering__schemas__PreviewRequest": {
        "additionalProperties": false,
        "properties": {
          "branding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandingUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "doc_type": {
            "default": "invoice",
            "enum": [
              "quote",
              "invoice",
              "receipt",
              "credit_note"
            ],
            "title": "Doc Type",
            "type": "string"
          },
          "format": {
            "default": "pdf",
            "enum": [
              "pdf",
              "html"
            ],
            "title": "Format",
            "type": "string"
          },
          "variant": {
            "default": "standard",
            "enum": [
              "standard",
              "long_names",
              "many_lines",
              "zero_dp",
              "rtl"
            ],
            "title": "Variant",
            "type": "string"
          }
        },
        "title": "PreviewRequest",
        "type": "object"
      },
      "app__modules__rendering__schemas__TemplateOut": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "doc_types": {
            "items": {
              "type": "string"
            },
            "title": "Doc Types",
            "type": "array"
          },
          "font_pair": {
            "title": "Font Pair",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "paper": {
            "title": "Paper",
            "type": "string"
          },
          "tier": {
            "enum": [
              "free",
              "premium"
            ],
            "title": "Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "name",
          "description",
          "tier",
          "version",
          "doc_types",
          "paper",
          "font_pair"
        ],
        "title": "TemplateOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BrokerOS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/branding": {
      "get": {
        "operationId": "branding_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Branding",
        "tags": [
          "branding"
        ]
      },
      "patch": {
        "operationId": "branding_update",
        "parameters": [
          {
            "in": "header",
            "name": "if-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Branding",
        "tags": [
          "branding"
        ]
      }
    },
...
    "/api/v1/templates": {
      "get": {
        "operationId": "templates_list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/app__modules__rendering__schemas__TemplateOut"
                  },
                  "title": "Response Templates List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Templates",
        "tags": [
          "branding"
        ]
      }
    },
    "/api/v1/templates/{key}/preview": {
      "post": {
        "description": "Render sample data with this template and the tenant's branding (or unsaved `branding` overrides).",
        "operationId": "templates_preview",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__modules__rendering__schemas__PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/pdf": {},
              "text/html": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview",
        "tags": [
          "branding"
        ]
      }
    },
...
  }
}